Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,40 +4,50 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
 	@Autowired
 	private UserRepository userRepository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
-	public User findById(Long id){
+
+	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if(userRepository.existsById(id))
+				userRepository.deleteById(id);
+			else
+				throw new ResourceNotFoundException(id);
+		}
+		catch(DataIntegrityViolationException ex) {
+			throw new DatabaseException(ex.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = userRepository.getReferenceById(id);
 		updateData(entity, obj);
 		return userRepository.save(entity);
 	}
-	
+
 	public void updateData(User entity, User obj) {
 		entity.setEmail(obj.getEmail());
 		entity.setName(obj.getName());
